% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_test.R
\name{autocorrelation_test}
\alias{autocorrelation_test}
\title{autocorrelation_test}
\usage{
autocorrelation_test(mod)
}
\arguments{
\item{mod}{is the linear model, for example, this can be the lm_surv, lm_lambda, or lm_abundance}
}
\description{
Examines autocorrelation in the regression models from the DemographicTrend package
}
\author{
Elizabeth Hiroyasu
}
