% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculating abundances with adjusted survivals
This function uses matrix multiplication to calculate abundances from
transition matrices and N0 data}
\usage{
calc_abundance(surv_trend, fert_rand, clon_rand, trans_mat, years)
}
\arguments{
\item{surv_trend}{calculated from trend function}

\item{trans_mat}{is an array of transition matrices from the COMPADRE database}

\item{years}{is the number of years in simulation}

\item{N0_rand}{a random matrix of 10 years of initial population data}

\item{fert_mat}{fertility matrix from COMPADRE database}
}
\description{
Calculating abundances with adjusted survivals
This function uses matrix multiplication to calculate abundances from
transition matrices and N0 data
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for 
Demographic Research (Germany). Available at www.compadre-db.org 
(data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
