% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pv.R
\name{calc_pv}
\alias{calc_pv}
\title{calc_pv}
\usage{
calc_pv(surv_mat, fert_mat, clon_mat, trans_mat, Nt, nstage, years, stage2mod,
  beta, active_stages, verbose = FALSE)
}
\arguments{
\item{surv_mat}{array of survival matrices from COMPADRE database}

\item{fert_mat}{array of fertility matrices from COMPADRE database}

\item{nstage}{number of stages in matrix data}

\item{years}{number of years in simulation}

\item{stage2mod}{stage that is modified by inserted survival trend}

\item{beta}{is the amount by which survival will be modified in a stage}
}
\description{
Calculates the p-values of survival and abundance regressions over time after a trend is inserted into survival.
}
\author{
Elizabeth Hiroyasu
}
