% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mat.R
\name{extract_mat}
\alias{extract_mat}
\title{extract_mat}
\usage{
extract_mat(MatrixData)
}
\arguments{
\item{MatrixData}{matrix population data from the COMPADRE matrix data}
}
\description{
Extracts survival matrices from the COMPADRE matrix data.
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for Demographic Research (Germany). Available at www.compadre-db.org (data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
