% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_active_stages.R
\name{filter_active_stages}
\alias{filter_active_stages}
\title{Filter Active Stages
This function filters matrix data to output matrices that have two or more active stages}
\usage{
filter_active_stages(tempMatClass, tempMatrixData)
}
\arguments{
\item{tempMatClass}{pre-filtered compadre data. To see other parameters the data is filtered by see the function Load_Compadre_Data.R}

\item{tempMatrixData}{prefiltered compadre matrices.}
}
\description{
Filter Active Stages
This function filters matrix data to output matrices that have two or more active stages
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for 
Demographic Research (Germany). Available at www.compadre-db.org 
(data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
