% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_survival_trend.R
\name{insert_survival_trend}
\alias{insert_survival_trend}
\title{Adding a trend to survival rates in different stages}
\usage{
insert_survival_trend(surv_rand, beta, nstage, stage2mod, years,
  verbose = FALSE)
}
\arguments{
\item{beta}{dictates how large the trend is to be introduced to survival matrices. A value of -beta will be added to survival each year}

\item{nstage}{is the number of stages represented in the matrix data}

\item{stage2mod}{the index of the stage for which survival is to be modified}

\item{years}{number of years in the simulation}

\item{surv_mat}{an array of randomized survival matrices}
}
\value{
surv_trend An array of matrices with a trend in survival

surv_zero a count of the number of times survival goes to zero after a trend is inserted
}
\description{
This function adds a trend to stage 2 survival
in the survival matrices for a stage or age structured population.
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for 
Demographic Research (Germany). Available at www.compadre-db.org 
(data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
