% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_regr.R
\name{lambda_regr}
\alias{lambda_regr}
\title{lambda_regr}
\usage{
lambda_regr(lambda, years)
}
\arguments{
\item{lambda}{is a 3d array of lambda matrices}

\item{years}{number of years in simulation}
}
\description{
Runs a regression analysis examining the power of detecting a trend in
modified lambda matrices
}
\author{
Elizabeth Hiroyasu
}
