% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_abundance_regr.R
\name{log_abundance_regr}
\alias{log_abundance_regr}
\title{log_abundance_regr}
\usage{
log_abundance_regr(abundance, years, active_stages)
}
\arguments{
\item{abundance}{array of abundance matrices calculated from transition matrices
with modified survival}

\item{years}{the number of years in the simulation}

\item{active_stages}{is the stages of the species that are active (as opposed to dormant or propagule)}
}
\description{
Runs a regression analysis examining the power of detecting a trend in
modified log abundance matrices
}
\author{
Elizabeth Hiroyasu
}
