% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pv.R
\name{plot_pv}
\alias{plot_pv}
\title{plot_pv}
\usage{
plot_pv(output, alpha)
}
\arguments{
\item{output}{is a list of different values, including survival p-values, abundance p-values, and lambda p-values}

\item{alpha}{is a vector of alpha values to compare p-values to and to assess power. Default alpha values are alpha=seq(from=0.01, to=0.2, by=delta)}
}
\description{
Plots different output values, and creates histograms of pvalues. This function can also be used to loop through a list of different outputs
}
\references{
data derived from COMPADRE Plant Matrix Database. Max Planck Institute for Demographic Research (Germany). Available at www.compadre-db.org (data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
