% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_lessthan_alpha.R
\name{pv_lessthan_alpha}
\alias{pv_lessthan_alpha}
\title{pv_lessthan_alpha}
\usage{
pv_lessthan_alpha(pvalue, alpha)
}
\arguments{
\item{pvalue}{is a vector of p-values}

\item{alpha}{is a vector of alpha values over which to test if abundance is less than}
}
\value{
prop_N a vector of proportional counts of the number of times the abundance p-value is less than the alpha value
}
\description{
counting the number of times p-value for abundance is less than alpha, this will 
allow us to assess power.
}
\references{
data derived from COMPADRE Plant Matrix Database. Max Planck Institute for Demographic Research (Germany). Available at www.compadre-db.org (data downloaded on [1/1/2014]).
}
\author{
Elizabeth Hiroyasu
}
