% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{Pval_coef}
\alias{Pval_coef}
\title{Extract a regression coefficient's P-value}
\usage{
Pval_coef(object, i_coef = 2, one.sided = TRUE, faster = TRUE)
}
\arguments{
\item{object}{A model object that supports \code{summary} and \code{coef}}

\item{i_coef}{Index of the coefficient for which the P-value is desired}

\item{one.sided}{Logical: should the P-value be a one-sided test? If so, negative coefficient estimates will generate P < 0.5}

\item{faster}{Logical: if true, use \code{\link{Pval_faster_lm}} to caluclate the P-value. Use this option only with an \code{object} of class \code{lm}}
}
\value{
The P-value of the specified coefficient
}
\description{
Extracts the P-value for a specified coefficient of a regresssion.
}
