% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{Pval_faster_lm}
\alias{Pval_faster_lm}
\title{Fast calculation of coefficient P-values}
\usage{
Pval_faster_lm(object)
}
\arguments{
\item{object}{A model object of class \code{lm}}
}
\value{
Vector of P-values, with same length as \code{coef(object)}
}
\description{
Calculates the coefficient P-values from an \code{lm} object. Uses identical code to \code{\link{summary.lm}}, without all the other calculations the other performs. No error checking is performed; outcomes with any other type of regression object are not guaranteed!
}
